changes for main_menu, sub_menu and SiteMap

+++ branches/0.8(django)/PyLucid/dev/make_install_data.py
@@ -12,19 +12,32 @@ sys.path[0] = os.getcwd()
 
 from PyLucid import settings
 
+
+
+SIMULATE = True
+APP_NAME = "PyLucid"
+
+
 # DB data files not needed for installation
 UNNEEDED_FILES = (
-    "archive", "md5user", "object_cache", "pages_internal", "plugin", "plugindata",
+    "archive", "md5user", "object_cache", "pages_internal", "plugin",
+    "plugindata",
 )
+
+
+
 
 def delete_file(path):
         print "delete '%s'" % filename,
-        os.remove(abs_path)
+        if SIMULATE:
+            print "[simulate only]",
+        else:
+            os.remove(abs_path)
         print "OK\n"
 
 filelist = os.listdir(settings.INSTALL_DATA_DIR)
 
-prefix_len = len(settings.TABLE_PREFIX)
+prefix_len = len(APP_NAME)
 filelist.sort()
 for filename in filelist:
     if filename.startswith("."):
@@ -33,7 +46,7 @@ for filename in filelist:
 
     abs_path = os.path.join(settings.INSTALL_DATA_DIR, filename)
 
-    if not filename.startswith(settings.TABLE_PREFIX):
+    if not filename.startswith(APP_NAME):
         # django tables
         delete_file(abs_path)
         continue
